+++ x/profiles/keeper/keeper_blocks.go
@@ -11,47 +11,36 @@ import (
 // something goes wrong.
 func (k Keeper) SaveUserBlock(ctx sdk.Context, userBlock types.UserBlock) error {
 	store := ctx.KVStore(k.storeKey)
-	key := types.UsersBlocksStoreKey(userBlock.Blocker)
-
-	blocks := types.MustUnmarshalUserBlocks(k.cdc, store.Get(key))
-	for _, ub := range blocks {
-		if ub == userBlock {
-			return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest,
-				"the user with address %s has already been blocked", userBlock.Blocked)
-		}
+	key := types.UsersBlocksStoreKey(userBlock.Blocker, userBlock.Subspace, userBlock.Blocked)
+	if store.Has(key) {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest,
+			"the user with address %s has already been blocked", userBlock.Blocked)
 	}
 
-	store.Set(key, types.MustMarshalUserBlocks(k.cdc, append(blocks, userBlock)))
+	store.Set(key, k.cdc.MustMarshalBinaryBare(&userBlock))
 	return nil
 }
 
 // DeleteUserBlock allows to the specified blocker to unblock the given blocked user.
 func (k Keeper) DeleteUserBlock(ctx sdk.Context, blocker, blocked string, subspace string) error {
 	store := ctx.KVStore(k.storeKey)
-	key := types.UsersBlocksStoreKey(blocker)
-
-	blocks := types.MustUnmarshalUserBlocks(k.cdc, store.Get(key))
-
-	blocks, found := types.RemoveUserBlock(blocks, blocker, blocked, subspace)
-	if !found {
+	key := types.UsersBlocksStoreKey(blocker, subspace, blocked)
+	if !store.Has(key) {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest,
 			"block from %s towards %s for subspace %s not found", blocker, blocked, subspace)
 	}
-
-	// Delete the key if no blocks are left.
-	// This cleans up the store avoiding export/import tests to fail due to a different number of keys present.
-	if len(blocks) == 0 {
-		store.Delete(key)
-	} else {
-		store.Set(key, types.MustMarshalUserBlocks(k.cdc, blocks))
-	}
+	store.Delete(key)
 	return nil
 }
 
 // GetUserBlocks returns the list of users that the specified user has blocked.
-func (k Keeper) GetUserBlocks(ctx sdk.Context, user string) []types.UserBlock {
-	store := ctx.KVStore(k.storeKey)
-	return types.MustUnmarshalUserBlocks(k.cdc, store.Get(types.UsersBlocksStoreKey(user)))
+func (k Keeper) GetUserBlocks(ctx sdk.Context, blocker string) []types.UserBlock {
+	var userblocks []types.UserBlock
+	k.IterateBlockedUsers(ctx, blocker, func(index int64, userblock types.UserBlock) (stop bool) {
+		userblocks = append(userblocks, userblock)
+		return false
+	})
+	return userblocks
 }
 
 // GetAllUsersBlocks returns a list of all the users blocks inside the given context.
@@ -63,8 +52,8 @@ func (k Keeper) GetAllUsersBlocks(ctx sdk.Context) []types.UserBlock {
 
 	var usersBlocks []types.UserBlock
 	for ; iterator.Valid(); iterator.Next() {
-		blocks := types.MustUnmarshalUserBlocks(k.cdc, iterator.Value())
-		usersBlocks = append(usersBlocks, blocks...)
+		block := types.MustUnmarshalUserBlock(k.cdc, iterator.Value())
+		usersBlocks = append(usersBlocks, block)
 	}
 
 	return usersBlocks
@@ -77,11 +66,17 @@ func (k Keeper) IsUserBlocked(ctx sdk.Context, blocker, blocked string) bool {
 
 // HasUserBlocked returns true if the provided blocker has blocked the given user for the given subspace.
 // If the provided subspace is empty, all subspaces will be checked
-func (k Keeper) HasUserBlocked(ctx sdk.Context, blocker, user, subspace string) bool {
-	blocks := k.GetUserBlocks(ctx, blocker)
+func (k Keeper) HasUserBlocked(ctx sdk.Context, blocker, blocked, subspace string) bool {
+	if subspace != "" {
+		store := ctx.KVStore(k.storeKey)
+		key := types.UsersBlocksStoreKey(blocker, subspace, blocked)
 
+		return store.Has(key)
+	}
+
+	blocks := k.GetUserBlocks(ctx, blocker)
 	for _, block := range blocks {
-		if block.Blocked == user {
+		if block.Blocked == blocked {
 			return subspace == "" || block.Subspace == subspace
 		}
 	}