+++ x/profiles/simulation/decoder_test.go
@@ -77,7 +77,7 @@ func TestDecodeStore(t *testing.T) {
 			Value: relBz,
 		},
 		{
-			Key:   types.UsersBlocksStoreKey(firstAddr),
+			Key:   types.UsersBlocksStoreKey(firstAddr, "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e", secondAddr),
 			Value: blocksBz,
 		},
 	}}